Move SIP_Validator out of Validations module.   Declutter the namespace.  Only validations are under Validations. Refactor spec to comply with namespace. Make Validator initialize more explicit.
