+++ lib/ht_sip_validator/validation.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 require "ht_sip_validator/validation/base"
-require "ht_sip_validator/validation/sip_validator"
+require "ht_sip_validator/sip_validator"
 require "ht_sip_validator/validation/meta_yml"
 require "ht_sip_validator/validation/message"
